requirements.txt
# Core Dependencies
PyPokerEngine==1.0.7

# LLM Service
openai>=1.0.0  # Deepseek 使用兼容 OpenAI 的 API
httpx>=0.24.0
aiohttp>=3.8.0

# CLI Interface
rich>=13.0.0
prompt_toolkit>=3.0.0

# Configuration
python-dotenv>=1.0.0

# Utils
pydantic>=2.0.0
pyyaml>=6.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0

# Development Tools
black>=23.0.0
mypy>=1.0.0
flake8>=6.0.0

# Logging
colorlog>=6.7.0


